Return component data as alists
Previously the lisp code was return `ComponentData` objects as
`Rc<dyn Any>`.  This is no longer the case and the alists returned by
the lisp code are converted to `ComponentData` on the rust side.

This allows writing tests in lisp for the component data methods,
because they just return alists and those can be verified by lisp.